ADD - Updated ECS

+++ src/ECS/Components.h
@@ -7,7 +7,7 @@ struct NameComponent {
 };
 
 struct TransformComponent {
-  glm::vec2 position;
+  glm::ivec2 position;
   glm::vec2 scale;
   double rotation;
 };
